pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vid2subs"
version = "0.1.0"
description = "Convert video or audio into time-aligned subtitles with optional vocal separation and translation."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [
    { name = "baicai1145" },
]
dependencies = [
    "torch==2.7.1",
    "torchaudio==2.7.1",
    "librosa>=0.10",
    "soundfile>=0.12",
    "pymss",
    "nemoasr2pytorch",
    "funasr",
    "requests>=2.31",
    "transformers>=4.40",
    "sentencepiece",
    "modelscope",
    "python-dotenv>=1.0",
]

[project.scripts]
vid2subs = "vid2subs.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
